Move Thorchain plugins into their own folder
