+++ test/thorchain.test.ts
@@ -1,7 +1,7 @@
 import { assert } from 'chai'
 import { describe, it } from 'mocha'
 
-import { getVolatilitySpread } from '../src/swap/defi/thorchain'
+import { getVolatilitySpread } from '../src/swap/defi/thorchain/thorchain'
 
 describe(`getVolatilitySpread`, function () {
   it('bitcoin source', function () {
